% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointsmsvoice_operations.R
\name{delete_configuration_set}
\alias{delete_configuration_set}
\title{Deletes an existing configuration set}
\usage{
delete_configuration_set(
  ConfigurationSetName = "string"
)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}
}
\description{
Deletes an existing configuration set.
}
\examples{

}
