% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointsmsvoice_operations.R
\name{delete_configuration_set_event_destination}
\alias{delete_configuration_set_event_destination}
\title{Deletes an event destination in a configuration set}
\usage{
delete_configuration_set_event_destination(
  ConfigurationSetName = "string",
  EventDestinationName = "string"
)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestinationName}{[required] EventDestinationName}
}
\description{
Deletes an event destination in a configuration set.
}
\examples{

}
