% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointsmsvoice_operations.R
\name{get_configuration_set_event_destinations}
\alias{get_configuration_set_event_destinations}
\title{Obtain information about an event destination, including the types of events it reports, the Amazon Resource Name (ARN) of the destination, and the name of the event destination}
\usage{
get_configuration_set_event_destinations(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}
}
\description{
Obtain information about an event destination, including the types of events it reports, the Amazon Resource Name (ARN) of the destination, and the name of the event destination.
}
\section{Accepted Parameters}{
\preformatted{get_configuration_set_event_destinations(
  ConfigurationSetName = "string"
)
}
}

