% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointsmsvoice_operations.R
\name{list_configuration_sets}
\alias{list_configuration_sets}
\title{List all of the configuration sets associated with your Amazon Pinpoint account in the current region}
\usage{
list_configuration_sets(NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to the API that indicates the position in the list of results.}

\item{PageSize}{Used to specify the number of items that should be returned in the response.}
}
\description{
List all of the configuration sets associated with your Amazon Pinpoint account in the current region.
}
\section{Accepted Parameters}{
\preformatted{list_configuration_sets(
  NextToken = "string",
  PageSize = "string"
)
}
}

