% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointsmsvoice_package.R
\docType{package}
\name{paws.pinpointsmsvoice-package}
\alias{paws.pinpointsmsvoice}
\alias{paws.pinpointsmsvoice-package}
\title{paws.pinpointsmsvoice: Amazon Pinpoint SMS and Voice Service}
\description{
Pinpoint SMS and Voice Messaging public facing APIs
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
